Add bargraph graph
