+++ DemoApp/Bedside_Monitor/Src/Graph_L1.c
@@ -1,0 +1,8 @@
+﻿static void graph_ecg() {}
+static void graph_ple() {}
+static void graph_co2() {}
+
+
+void graph_l1() {
+
+}
